split tests into separate files test run.
