+++ bin/get_loca_files.sh
