--- nohup.out
@@ -1,55 +1,0 @@
-· Creating environments
-· Discovering benchmarks
-·· Uninstalling from conda-py3.6-dask-distributed-gcsfs-h5netcdf-h5py-lz4-netcdf4-numpy-pip+dask_kubernetes-pip+h5pyd-pyyaml-rasterio-xarray-zarr
-·· Building for conda-py3.6-dask-distributed-gcsfs-h5netcdf-h5py-lz4-netcdf4-numpy-pip+dask_kubernetes-pip+h5pyd-pyyaml-rasterio-xarray-zarr
-·· Installing into conda-py3.6-dask-distributed-gcsfs-h5netcdf-h5py-lz4-netcdf4-numpy-pip+dask_kubernetes-pip+h5pyd-pyyaml-rasterio-xarray-zarr
-·· Error running /home/jovyan/dev/storage-benchmarks/.asv/env/4147cb8bacd14ebb1c15e8248dc80f49/bin/python /home/jovyan/.local/lib/python3.6/site-packages/asv/benchmark.py discover /home/jovyan/dev/storage-benchmarks/benchmarks /tmp/tmpy4gfkt7n
-             STDOUT -------->
-             
-             STDERR -------->
-             Traceback (most recent call last):
-               File "/home/jovyan/.local/lib/python3.6/site-packages/asv/benchmark.py", line 795, in <module>
-                 commands[mode](args)
-               File "/home/jovyan/.local/lib/python3.6/site-packages/asv/benchmark.py", line 740, in main_discover
-                 list_benchmarks(benchmark_dir, fp)
-               File "/home/jovyan/.local/lib/python3.6/site-packages/asv/benchmark.py", line 725, in list_benchmarks
-                 for benchmark in disc_benchmarks(root):
-               File "/home/jovyan/.local/lib/python3.6/site-packages/asv/benchmark.py", line 710, in disc_benchmarks
-                 for module in disc_files(root, os.path.basename(root) + '.'):
-               File "/home/jovyan/.local/lib/python3.6/site-packages/asv/benchmark.py", line 699, in disc_files
-                 module = import_module(package + filename)
-               File "/home/jovyan/dev/storage-benchmarks/.asv/env/4147cb8bacd14ebb1c15e8248dc80f49/lib/python3.6/importlib/__init__.py", line 126, in import_module
-                 return _bootstrap._gcd_import(name[level:], package, level)
-               File "<frozen importlib._bootstrap>", line 994, in _gcd_import
-               File "<frozen importlib._bootstrap>", line 971, in _find_and_load
-               File "<frozen importlib._bootstrap>", line 955, in _find_and_load_unlocked
-               File "<frozen importlib._bootstrap>", line 665, in _load_unlocked
-               File "<frozen importlib._bootstrap_external>", line 678, in exec_module
-               File "<frozen importlib._bootstrap>", line 219, in _call_with_frames_removed
-               File "/home/jovyan/dev/storage-benchmarks/benchmarks/target_hsds.py", line 7, in <module>
-                 import h5pyd
-             ModuleNotFoundError: No module named 'h5pyd'
-Traceback (most recent call last):
-  File "/home/jovyan/.local/bin//asv", line 11, in <module>
-    sys.exit(main())
-  File "/home/jovyan/.local/lib/python3.6/site-packages/asv/main.py", line 38, in main
-    result = args.func(args)
-  File "/home/jovyan/.local/lib/python3.6/site-packages/asv/commands/__init__.py", line 48, in run_from_args
-    return cls.run_from_conf_args(conf, args)
-  File "/home/jovyan/.local/lib/python3.6/site-packages/asv/commands/run.py", line 127, in run_from_conf_args
-    **kwargs
-  File "/home/jovyan/.local/lib/python3.6/site-packages/asv/commands/run.py", line 185, in run
-    benchmarks = Benchmarks(conf, repo, environments, regex=bench)
-  File "/home/jovyan/.local/lib/python3.6/site-packages/asv/benchmarks.py", line 299, in __init__
-    benchmarks = self.disc_benchmarks(conf, repo, environments)
-  File "/home/jovyan/.local/lib/python3.6/site-packages/asv/benchmarks.py", line 348, in disc_benchmarks
-    dots=False)
-  File "/home/jovyan/.local/lib/python3.6/site-packages/asv/plugins/conda.py", line 189, in run
-    return self.run_executable('python', args, **kwargs)
-  File "/home/jovyan/.local/lib/python3.6/site-packages/asv/plugins/conda.py", line 195, in run_executable
-    return super(Conda, self).run_executable(executable, args, **kwargs)
-  File "/home/jovyan/.local/lib/python3.6/site-packages/asv/environment.py", line 581, in run_executable
-    return util.check_output([exe] + args, **kwargs)
-  File "/home/jovyan/.local/lib/python3.6/site-packages/asv/util.py", line 526, in check_output
-    raise ProcessError(args, retcode, stdout, stderr)
-asv.util.ProcessError: Command '/home/jovyan/dev/storage-benchmarks/.asv/env/4147cb8bacd14ebb1c15e8248dc80f49/bin/python /home/jovyan/.local/lib/python3.6/site-packages/asv/benchmark.py discover /home/jovyan/dev/storage-benchmarks/benchmarks /tmp/tmpy4gfkt7n' returned non-zero exit status 1
