+++ results/kubernetes-gcp/45c0ba24-conda-py3.6-dask-distributed-gcsfs-h5netcdf-h5py-lz4-netcdf4-numpy-pip+dask_kubernetes-pip+h5pyd-pyyaml-rasterio-xarray-zarr.json
@@ -2,8 +2,10 @@
     "commit_hash": "45c0ba242945972cac6e929ac66968a04f322210",
     "date": 1529112143000,
     "ended_at": {
-        "gcp_zarr_dask_read.Report_dataset_sizes.track_Zarr_GCP_LLC4320_dataset_size": 1532021218000,
-        "gcp_zarr_dask_read.Zarr_GCP_LLC4320.time_load_array_compute_theta_mean": 1532028017000
+        "gcp_kubernetes_read_netCDF_LLC4320.Theta_max.time_read": 1532353946000,
+        "gcp_kubernetes_read_netCDF_LLC4320.report_ds_mbytes.track_gcp_llc4320_netcdf_10": 1532353949000,
+        "gcp_kubernetes_read_zarr_LLC4320.Theta_max.time_read": 1532354093000,
+        "gcp_kubernetes_read_zarr_LLC4320.report_ds_mbytes.track_gcp_llc4320_zarr_10": 1532354096000
     },
     "env_name": "conda-py3.6-dask-distributed-gcsfs-h5netcdf-h5py-lz4-netcdf4-numpy-pip+dask_kubernetes-pip+h5pyd-pyyaml-rasterio-xarray-zarr",
     "params": {
@@ -47,23 +49,7 @@
         "zarr": ""
     },
     "results": {
-        "gcp_zarr_dask_read.Report_dataset_sizes.track_Zarr_GCP_LLC4320_dataset_size": {
-            "params": [
-                [
-                    "'ALL'"
-                ],
-                [
-                    "1"
-                ],
-                [
-                    "1"
-                ]
-            ],
-            "result": [
-                1085.5994440540671
-            ]
-        },
-        "gcp_zarr_dask_read.Zarr_GCP_LLC4320.time_load_array_compute_theta_mean": {
+        "gcp_kubernetes_read_netCDF_LLC4320.Theta_max.time_read": {
             "params": [
                 [
                     "'GCS'"
@@ -83,30 +69,57 @@
                 ]
             ],
             "result": [
-                337.15150225700927,
-                339.4921992960153,
-                190.59802098499495,
-                179.6019520520058,
-                160.0696038820024,
-                145.3084786820109,
-                296.084964337002,
-                230.27593282301677,
-                198.13949139599572,
-                171.82577124598902,
-                157.71726426298846,
-                141.5535019270028,
-                297.6882702719886,
-                236.25114229600877,
-                196.99455105501693,
-                164.0614375229925,
-                171.07648325400078,
-                154.55136649601627
+                46.97942448593676,
+                36.26193560601678,
+                31.347557924920693,
+                27.814826411078684,
+                24.898538035922684,
+                22.442745489999652,
+                12.729238392901607,
+                11.874191293027252,
+                12.11855266103521,
+                12.642860352061689,
+                13.723776725004427,
+                11.548531744047068,
+                12.4363614609465,
+                12.205289943958633,
+                12.310305322986096,
+                12.153349547996186,
+                12.258691405993886,
+                12.177306066034362
             ]
-        }
+        },
+        "gcp_kubernetes_read_netCDF_LLC4320.report_ds_mbytes.track_gcp_llc4320_netcdf_10": 11212.663360595703,
+        "gcp_kubernetes_read_zarr_LLC4320.Theta_max.time_read": {
+            "params": [
+                [
+                    "'GCS'"
+                ],
+                [
+                    "60",
+                    "80",
+                    "100",
+                    "120",
+                    "140",
+                    "160"
+                ]
+            ],
+            "result": [
+                6.742074011010118,
+                5.237266714917496,
+                8.480406768037938,
+                8.512810976011679,
+                8.713658352964558,
+                8.605634883977473
+            ]
+        },
+        "gcp_kubernetes_read_zarr_LLC4320.report_ds_mbytes.track_gcp_llc4320_zarr_10": 11212.663360595703
     },
     "started_at": {
-        "gcp_zarr_dask_read.Report_dataset_sizes.track_Zarr_GCP_LLC4320_dataset_size": 1532021200000,
-        "gcp_zarr_dask_read.Zarr_GCP_LLC4320.time_load_array_compute_theta_mean": 1532021218000
+        "gcp_kubernetes_read_netCDF_LLC4320.Theta_max.time_read": 1532352022000,
+        "gcp_kubernetes_read_netCDF_LLC4320.report_ds_mbytes.track_gcp_llc4320_netcdf_10": 1532353946000,
+        "gcp_kubernetes_read_zarr_LLC4320.Theta_max.time_read": 1532353949000,
+        "gcp_kubernetes_read_zarr_LLC4320.report_ds_mbytes.track_gcp_llc4320_zarr_10": 1532354093000
     },
     "version": 1
 }
